// src/models/common/models.enum.ts
export enum Roles {
  SuperAdmin = "SuperAdmin",
  Admin = "Admin",
  Users = "Users",
  Anonymous = "Anonymous",
}

export enum CategoryLevel {
  Level1 = "1",
  Level2 = "2",
  Level3 = "3",
}

export enum ObjectStatus {
  Active = "active",
  DeActive = "deactive",
}

export enum OrderDirection {
  DESC = "DESC",
  ASC = "ASC",
}

export enum CashType {
  Points = "Points",
  Prices = "Prices",
}
export enum PaidType {
  Cash = "Cash",
  Points = "Points",
  Momo = "Momo",
  InternationalCard = "InternationalCard",
  ATM = "ATM",
}

export enum FileType {
  Product = "Product",
  ProductThumbnail = "ProductThumbnail",
  UserAvatar = "UserAvatar",
  Articles = "Articles",
  Banner = "Banner",
  Challenges = "Challenges",
  Gift = "Gift",
  DiscountsThumb = "DiscountsThumb",
  DiscountsAvatar = "DiscountsAvatar",
}

export enum ActionType {
  Edit = "Edit",
  Create = "Create",
  Delete = "Delete",
}

export enum ShippingStatus {
  Create = "Create",
  Cash = "Cash",
  Pack = "Pack",
  Delivery = "Delivery",
  Delivered = "Delivered",
  Refund = "Refund",
  Refunded = "Refunded",
}